# 3D plotting performance improvements