--- a/doc/users/next_whats_new/3d_speedups.rst
+++ b/doc/users/next_whats_new/3d_speedups.rst
@@ -1,0 +1,6 @@
+3D performance improvements
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Draw time for 3D plots has been improved, especially for surface and wireframe
+plots. Users should see up to a 10x speedup in some cases. This should make
+interacting with 3D plots much more responsive.