… lib/mpl_toolkits/mplot3d/tests/baseline_images/test_axes3d/surface3d_masked.png