## PR summary
This builds off of https://github.com/matplotlib/matplotlib/pull/16688, to rebase off of main, extend as applicable to all the 3D object types, and allow for compatibility with the new 3D axlim clipping. With the test code below, I am seeing substantial speedups in draw times for some types of plots.

Closes https://github.com/matplotlib/matplotlib/issues/16659

```python
import time
import numpy as np
import matplotlib.pyplot as plt

fig, ax = plt.subplots(subplot_kw={"projection": "3d"})

x = y = np.linspace(-1, 1, 100)
X, Y = np.meshgrid(x, y)
Z = X ** 2 + Y ** 2

print("Timing...")

start_time = time.perf_counter()
#ax.plot_surface(X, Y, Z, rstride=1, cstride=1, axlim_clip=False)
#ax.scatter(x, y, np.ones_like(x), axlim_clip=False)
ax.plot_wireframe(X, Y, Z, axlim_clip=False)
#ax.bar3d(X.ravel(), Y.ravel(), 0, 0.01, 0.01, 1, axlim_clip=False)
ax.set_xlim(0, 1)
fig.canvas.draw()
end_time = time.perf_counter()

plt.close()
print(f"Time taken: {end_time - start_time:.4f} seconds")
```

| function                   | object            | draw time main | this PR  | improvement |
|----------------------------|-------------------|----------|----------|-------------|
| plot_surface (400x400)     | Poly3DCollection  | 16.55 sec | 1.86 sec  | 8.9x        |
| scatter (100000)           | Path3DCollection  | 3.65 sec | 3.22 sec  | 1.1x        |
| plot wireframe (8000x8000) | Line3DCollection  | 2.72 sec | 1.39 sec | 2.0x        |
| bard3d (100x100)           | Patch3DCollection | 7.17 sec | 2.02 sec | 3.5x        |

## PR checklist

- [n/a] "closes #0000" is in the body of the PR description to [link the related issue](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue)
- [x] new and changed code is [tested](https://matplotlib.org/devdocs/devel/testing.html)
- [n/a] *Plotting related* features are demonstrated in an [example](https://matplotlib.org/devdocs/devel/document.html#write-examples-and-tutorials)
- [n/a] *New Features* and *API Changes* are noted with a [directive and release note](https://matplotlib.org/devdocs/devel/api_changes.html#announce-changes-deprecations-and-new-features)
- [x] Documentation complies with [general](https://matplotlib.org/devdocs/devel/document.html#write-rest-pages) and [docstring](https://matplotlib.org/devdocs/devel/document.html#write-docstrings) guidelines