@scottshambaugh 2025-01-05
Thoughts on if this speedup warrants a what's new?