@timhoffm 2025-01-06
> Thoughts on if the 3D speedups warrant a what's new?

Since these are substantial, I‘d say yes.